pyproject.toml
[project]
name = "mow"
version = "0.1.0"
description = "Mow - a media organization workflow tool"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "folium>=0.17.0",
    "gpxpy>=1.6.2",
    "pillow>=11.0.0",
    "polars>=1.10.0",
    "pyexiftool>=0.5.6",
    "pyyaml>=6.0.2",
    "rich>=13.9.2",
]


[tool.pytest.ini_options]
log_cli = false
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
